requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
dbt-core==1.8.7
dbt-postgres==1.8.2
